requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
scipy>=1.7.0
pandas>=1.3.0
pygame>=2.0.0
jsonschema>=4.0.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.0
pillow>=8.0.0
pytest>=6.0.0
pytest-cov>=2.0.0
black>=22.0.0
isort>=5.0.0
